Fixed column type

+++ backend/migrations/20260505000001_add_geo_fields.sql
@@ -5,6 +5,6 @@ ALTER TABLE devices
   ADD COLUMN IF NOT EXISTS geo_display      TEXT;
 
 CREATE TABLE IF NOT EXISTS geo_update_queue (
-  device_id INTEGER PRIMARY KEY REFERENCES devices(device_id) ON DELETE CASCADE,
+  device_id BIGINT PRIMARY KEY REFERENCES devices(device_id) ON DELETE CASCADE,
   queued_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
 );
